@@ -27,57 +27,59 @@
     "helmet": "^3.21.2",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.15",
-    "mongodb": "^3.5.0",
-    "mongoose": "^5.8.7",
+    "mongodb": "^3.5.1",
+    "mongoose": "^5.8.9",
     "prop-types": "^15.7.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "react-ga": "^2.7.0",
-    "react-redux": "^5.1.2",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "redux": "^3.6.0",
+    "react-redux": "^7.1.3",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "redux": "^4.0.5",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.8.3",
+    "@babel/preset-env": "^7.8.3",
+    "@babel/preset-react": "^7.8.3",
     "babel": "^6.23.0",
     "babel-core": "^6.26.3",
-    "babel-eslint": "^8.2.6",
-    "babel-loader": "^7.1.5",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^8.0.6",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
     "chai": "^4.2.0",
     "chai-http": "^4.3.0",
-    "css-loader": "^0.28.11",
-    "eslint": "^4.19.1",
+    "css-loader": "^3.4.2",
+    "eslint": "^6.8.0",
     "eslint-plugin-react": "^7.18.0",
-    "eslint-watch": "^3.1.5",
+    "eslint-watch": "^6.0.1",
     "html-webpack-plugin": "^3.2.0",
     "ignore-styles": "^5.0.1",
-    "jsdom": "^11.12.0",
+    "jsdom": "^16.0.1",
     "json-loader": "^0.5.7",
     "mini-css-extract-plugin": "^0.9.0",
-    "mocha": "5.1.1",
+    "mocha": "^7.0.0",
     "nodemon": "^2.0.2",
     "npm-run-all": "^4.1.5",
-    "optimize-css-assets-webpack-plugin": "3.2.0",
-    "style-loader": "^0.19.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.3",
+    "style-loader": "^1.1.3",
     "terser-webpack-plugin": "^2.3.2",
-    "url-loader": "^1.1.2",
+    "url-loader": "^3.0.0",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
-    "webpack-dev-middleware": "^1.12.2",
+    "webpack-dev-middleware": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0",
     "webpack-md5-hash": "^0.0.6",
     "yarn": "^1.21.1"
   },
   "babel": {
     "presets": [
-      "react",
-      "es2015",
-      "stage-0"
+      "@babel/preset-env",
+      "@babel/preset-react"
     ]
   },
   "keywords": [
