@@ -1,3 +1,3 @@
 /* eslint func-names: off, no-empty-function: off */
-require('babel-core/register')();
-require.extensions['.css'] = function() {};
+// require('babel-core/register')();
+// require.extensions['.css'] = function() {};
